requirements.txt
playwright
schedule
requests
python-dotenv
